requirements.txt
amqp==2.4.1
anyjson==0.3.3
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
billiard==3.5.0.5
celery==4.2.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
cycler==0.10.0
Django==2.1.7
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-haystack==2.8.1
django-timezone-field==3.0
dnspython==1.16.0
dukpy==0.2.2
eventlet==0.24.1
greenlet==0.4.15
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
jieba==0.39
Jinja2==2.10
jupyter-echarts-pypkg==0.1.2
kiwisolver==1.0.1
kombu==4.3.0
lml==0.0.2
lxml==4.3.2
MarkupSafe==1.1.1
monotonic==1.5
mysql==0.0.2
mysqlclient==1.4.2.post1
numpy==1.16.2
parsel==1.5.1
Pillow==5.4.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
pyecharts-jupyter-installer==0.0.3
PyHamcrest==1.9.0
PyMySQL==0.9.3
pyOpenSSL==18.0.0
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2018.9
queuelib==1.5.0
requests==2.21.0
service-identity==17.0.0
six==1.11.0
tqdm==4.26.0
urllib3==1.24.1
vine==1.2.0
virtualenv==16.0.0
w3lib==1.19.0
Whoosh==2.7.4
zope.interface==4.6.0
